buffer_size is in buffer_traits.hpp
