+++ include/boost/beast/_experimental/test/impl/stream.hpp
@@ -11,9 +11,10 @@
 #define BOOST_BEAST_TEST_IMPL_STREAM_HPP
 
 #include <boost/beast/core/bind_handler.hpp>
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/detail/service_base.hpp>
+#include <boost/beast/core/detail/type_traits.hpp>
 #include <mutex>
 #include <stdexcept>
 #include <vector>
