+++ include/boost/beast/core.hpp
@@ -15,7 +15,6 @@
 #include <boost/beast/core/async_base.hpp>
 #include <boost/beast/core/basic_stream.hpp>
 #include <boost/beast/core/bind_handler.hpp>
-#include <boost/beast/core/buffer_size.hpp>
 #include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffered_read_stream.hpp>
 #include <boost/beast/core/buffers_adaptor.hpp>
