+++ include/boost/beast/core/impl/basic_stream.hpp
@@ -11,7 +11,7 @@
 #define BOOST_BEAST_CORE_IMPL_BASIC_STREAM_HPP
 
 #include <boost/beast/core/async_base.hpp>
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/detail/type_traits.hpp>
 #include <boost/beast/websocket/teardown.hpp>
