+++ include/boost/beast/core/impl/buffers_adaptor.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_BEAST_IMPL_BUFFERS_ADAPTOR_HPP
 #define BOOST_BEAST_IMPL_BUFFERS_ADAPTOR_HPP
 
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/detail/type_traits.hpp>
 #include <boost/asio/buffer.hpp>
 #include <boost/config/workaround.hpp>
