+++ include/boost/beast/core/make_printable.hpp
@@ -11,7 +11,7 @@
 #define BOOST_BEAST_MAKE_PRINTABLE_HPP
 
 #include <boost/beast/core/detail/config.hpp>
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/asio/buffer.hpp>
 #include <ostream>
