+++ include/boost/beast/http/basic_dynamic_body.hpp
@@ -11,7 +11,7 @@
 #define BOOST_BEAST_HTTP_BASIC_DYNAMIC_BODY_HPP
 
 #include <boost/beast/core/detail/config.hpp>
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/detail/buffer.hpp>
 #include <boost/beast/http/error.hpp>
 #include <boost/beast/http/message.hpp>
