+++ test/beast/core/Jamfile
@@ -22,7 +22,6 @@ local SOURCES =
     async_base.cpp
     basic_stream.cpp
     bind_handler.cpp
-    buffer_size.cpp
     buffer_traits.cpp
     buffered_read_stream.cpp
     buffers_adapter.cpp
