+++ test/beast/core/buffers_prefix.cpp
@@ -12,7 +12,7 @@
 
 #include "test_buffer.hpp"
 
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffers_to_string.hpp>
 #include <boost/beast/_experimental/unit_test/suite.hpp>
 #include <string>
