+++ test/beast/http/file_body.cpp
@@ -10,7 +10,7 @@
 // Test that header file is self-contained.
 #include <boost/beast/http/file_body.hpp>
 
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/file_stdio.hpp>
 #include <boost/beast/core/flat_buffer.hpp>
