+++ test/beast/websocket/test.hpp
@@ -11,7 +11,7 @@
 #define BEAST_TEST_WEBSOCKET_TEST_HPP
 
 #include <boost/beast/core/bind_handler.hpp>
-#include <boost/beast/core/buffer_size.hpp>
+#include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/buffers_to_string.hpp>
 #include <boost/beast/core/ostream.hpp>
